Apply code review suggestions
