+++ base.d.ts
@@ -32,7 +32,7 @@ export {Entry} from './source/entry';
 export {Entries} from './source/entries';
 export {SetReturnType} from './source/set-return-type';
 export {Asyncify} from './source/asyncify';
-export {TakeLast} from './source/take-last';
+export {LastArrayElement} from './source/last-array-element';
 
 // Miscellaneous
 export {PackageJson} from './source/package-json';
