+++ readme.md
@@ -87,7 +87,7 @@ Click the type names for complete docs.
 - [`Entries`](source/entries.d.ts) - Create a type that represents the type of the entries of a collection.
 - [`SetReturnType`](source/set-return-type.d.ts) - Create a function type with a return type of your choice and the same parameters as the given function type.
 - [`Asyncify`](source/asyncify.d.ts) - Create an async version of the given function type.
-- [`TakeLast`](ts41/take-last.d.ts) - Extracts the type of the last element of an array.
+- [`LastArrayElement`](ts41/last-array-element.d.ts) - Extracts the type of the last element of an array.
 
 ### Template literal types
 
@@ -98,7 +98,7 @@ Click the type names for complete docs.
 - [`PascalCase`](ts41/pascal-case.d.ts) – Converts a string literal to pascal-case (`FooBar`)
 - [`SnakeCase`](ts41/snake-case.d.ts) – Convert a string literal to snake-case (`foo_bar`).
 - [`DelimiterCase`](ts41/delimiter-case.d.ts) – Convert a string literal to a custom string delimiter casing.
-- [`QueryResult`](ts41/query-result.d.ts) - Matches element types returned from a query.
+- [`QueryResult`](ts41/query-result.d.ts) - Matches element types returned from an HTML element query selector.
 - [`Split`](ts41/split.d.ts) - Represents an array of strings split using a passed-in character or character set.
 - [`Trim`](ts41/trim.d.ts) - Removes spaces from the edges of a string.
