+++ ts41/split.d.ts
@@ -4,6 +4,9 @@ Represents an array of strings split using a passed-in character or character se
 Use case: defining the return type of the `String.prototype.split` method.
 
 @example
+```
+import {Split} from 'type-fest';
+
 declare function split<S extends string, D extends string>(string: S, separator: D): Split<S, D>;
 
 type Item = 'foo' | 'bar' | 'baz' | 'waldo';
@@ -11,6 +14,7 @@ const items = 'foo,bar,baz,waldo';
 let array: Item[];
 
 array = split(items, ',');
+```
 */
 export type Split<S extends string, D extends string> =
     S extends `${infer T}${D}${infer U}`
